Fix config file for database
